// frontend/src/api/generated/models/GraphEdge.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Edge in knowledge graph.
 */
export type GraphEdge = {
    id: number;
    source: string;
    target: string;
    type: string;
    strength: number;
    label: (string | null);
};